moved Funds logic from router to FundsController
